Improve diffing children that have changed
We took a very simplistic view previously, that meant that any child that was different
was not "similar".  This worked for cases where a child was removed or inserted, but failed
when a child was changed (the output suggested to remove the actual child, and insert the 
expected child, rather than making the changes to the actual).

We now try both approaches if there are both inserts and removals, and check which has 
the lowest weight.