Fix bug with child elements with ignored options
The original code would treat a non-exact match as "different", but would then not add 
the diff, but use the weight. This was an old hangover from old code that needed the 
difference in order to render it as a diff, such that wrappers would get rendered correctly.
This is no longer needed.

This means that we no longer need the real/total weights. However, I'm not yet 
convinced that we will never need them, (or a client will never need them, even though
they aren't exposed yet), so, I'm leaving them in for now.