Add diffing for class / className
The adapter needs to identify the class attribute name - if the actual
and expected class attribute name match, then extended class diffing can
be done. Three new flags are introduced:

diffExactClasses:  true (default) to treat class as a standard
attribute, and diff normally as a string (i.e. order / spacing matters)
When false, then will diff it intelligently, ignoring spacing and order.
Other options mean optionally extra / missing classes can be ignored
too.

diffExtraClasses: when true, extra classes in the actual class are
treated as a difference. When false, they are ignored

diffMissingClasses: when true, missing classes in the actual class are
treated as a difference. When false, they are ignored