Convert tests to use different Expected/Actual adapters
A bug has come up (bruderstein/unexpected-react#9) that shows an issue where an 
"actual" object is passed to an "expected" adapter. This is due to array-changes using
the "actual" value for an "insert" operation if the element in the array has moved.

In order to catch future problems like this, the tests here have been converted to
throw an error if an "actual" value is ever passed to an "expected" adapter, or vice versa.

Note that the actual bug is *not* fixed in this commit. This is merely the preperation 
to allow that tests be written. Sadly, all the tests still worked, so we wouldn't have 
picked up on this anyway.