render() / Painter now take `diff` and `inspect` functions
This ensures the correct output format is used for tests.
Using expect.diff and expect.inspect means the currently selected output format is chosen, not that that should be used to render the current output.
Now that we have a separate render() function, it's no problem to pass the diff and inspect functions in.